waiting queue implementation purely using atomics
