+++ CMakeLists.txt
@@ -949,8 +949,8 @@ set(LIBNETDATA_FILES
         src/libnetdata/locks/rw-spinlock.h
         src/libnetdata/atomics/atomic_flags.h
         src/libnetdata/atomics/atomics.h
-        src/libnetdata/waiting-queue/waiting-queue.c
-        src/libnetdata/waiting-queue/waiting-queue.h
+        src/libnetdata/locks/waitq.c
+        src/libnetdata/locks/waitq.h
         src/libnetdata/object-state/object-state.c
         src/libnetdata/object-state/object-state.h
 )
